pyproject.toml
[tool.poetry]
name = "droplet"
version = "0.1.0"
description = ""
authors = ["Zhaka <shingis.zhansen@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}

[tool.poetry.group.bridge.dependencies]
pyserial = "^3.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
